[chore] 회원가입시 사용되는 서비스 로직 변경
