+++ client/src/service/user/index.ts
@@ -1,0 +1,5 @@
+import createUser from './createUser';
+
+import checkDuplicateUser from './checkDuplicateUser';
+
+export const userManager = { createUser, checkDuplicateUser };
